Introduce an external handle and an internal state
We are introducing an internal state, which is handled by a private
struct. This allows us to keep implementation specific details to the
library and force callees to go through the public interface.

We are using a new struct XiHandle to the user to identify a specific
instance, which can be created using xi_init().