lib: use xi_rpc::RpcLoop::embedded_mainloop() and xi_rpc::RpcLoop::send_message
Our referenced version of xi-editor contains a specialized event loop allowing for easy embedding. We are now using this method. embedded_mainloop communicates with the embedder through a thread safe deque. xi_rpc::RpcLoop::send_message pushes messages to the deque. We will implement a receiving loop, taking messages from the deque and handing it to the embedder's callback.